% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-pkg.R
\docType{package}
\name{annoFuseData-pkg}
\alias{annoFuseData-pkg}
\title{annoFuseData}
\description{
\code{annoFuseData} is a Bioconductor package that provides ...
}
\details{
handle imports here if necessary
}
