% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{exonsToPlot}
\alias{exonsToPlot}
\title{exonsToPlot.RDS}
\format{
When read in via readRDS, the following information is available:
\itemize{
\item contig: Denotes chromosome name
\item type: Denotes if coordinates correspond to gene or exon
\item start: Denotes genomic start position
\item end: Denotes genomic end position
\item strand: Denotes strand the feature is on
\item attributes: Feature attributes in from gtf file
\item geneName: Denotes gene symbol
\item geneID: Denotes Ensemble gene id
\item transcript: Denotes Ensemble transcript id
\item exonNumber: Denotes exon rank in transcript
}
}
\description{
detailed desc
58347 observations are included, with details about 7 different features (
see details below).
}
\references{
Feature information extracted from gencode.v27.primary_assembly.annotation.gtf.gz
}
\keyword{datasets}
