% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{expr_collapsed}
\alias{expr_collapsed}
\title{expr_collapsed.rds}
\format{
When read in via readRDS, the following information is available:
\itemize{
\item gene_id: Denotes gene symbol
\item BS_W97QQYKQ: Denotes sample id for example runs
}
}
\description{
detailed desc
58581 * 6 columns
}
\references{
if any
}
\keyword{datasets}
