% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{final_fusion_mmc2}
\alias{final_fusion_mmc2}
\title{final_fusion_mmc2.txt}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item Cancer: Denotes TCGA cancer types
\item Sample: Denotes TCGA sample id
\item Fusion: Denotes TCGA fusion call
\item Junction: Denotes Number of supporting split read supporting the site of putative fusion
\item Spanning: Number of spanning fragments suporting the site
\item Breakpoint1: Genomic coordinate of 5` breakpoint
\item Breakpoint2: Genomic coordinate of 3` breakpoint
}
}
\description{
detailed desc
25664 * 7cols
}
\references{
if any
}
\keyword{datasets}
