% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{fusionreference}
\alias{fusionreference}
\title{fusionreference.txt}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item FusionName: Denotes Fusion name
\item file: Source file used for annotation
\item type: Denotes annotation type; values are TCGAFusion
}
}
\description{
detailed desc
17754 * 3
}
\keyword{datasets}
