% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{merged_starfusion_tcga}
\alias{merged_starfusion_tcga}
\title{merged_starfusion_tcga.tsv}
\format{
When read in via read.delim(...,check.names=FALSE,stringsAsFactors = F), the following information is available:
\itemize{
\item #FusionName: Denotes Fusion name
\item JunctionReadCount: Number of supporting split read supporting the site of putative fusion
\item SpanningFragCount: Number of spanning fragments suporting the site
\item SpliceType: Denotes whether breakpoint is located on an exon junction
\item LeftGene: 5` end of fusion
\item LeftBreakpoint: Genomic coordinate of 5` breakpoint
\item RightGene: 3` end of fusion
\item RightBreakpoint: Genomic coordinate of 3` breakpoint
\item LargeAnchorSupport: Denotes whether there are split reads that provide long anchor supporting read on both sides of breakpoint
\item FFPM; Denotes exprression estimate
\item LeftBreakDinuc: Denotes 2 bases flanking to the left of the breakpoint
\item LeftBreakEntropy: Denotes Shannon Entropy for 15 exon bases flanking to the left of the breakpoint
\item RightBreakDinuc: Denotes 2 bases flanking to the right of the breakpoint
\item RightBreakEntropy:Denotes Shannon Entropy for 15 exon bases flanking to the right of the breakpoint
\item annots: Annotation from FusionAnnotator
\item CDS_LEFT_ID: Denotes cds_id flanking to the left of the breakpoint
\item CDS_LEFT_RANGE: Denotes cds range to the left of the breakpoint
\item CDS_RIGHT_ID: Denotes cds_id flanking to the right of the breakpoint
\item CDS_RIGHT_RANGE: Denotes cds range to the right of the breakpoint
\item PROT_FUSION_TYPE: Denotes if the 3` gene in fusion is fused either as INFRAME, FRAMESHIFT or other
\item FUSION_MODEL: Denotes gene structure of the reconstructed fusion
\item FUSION_CDS: Denotes cds sequence of the reconstructed fusion
\item FUSION_TRANSL: Denotes translated sequence of the reconstructed fusion
\item PFAM_LEFT: Denotes pfam id and location for reconstructed fusion transcript left to the breakpoint
\item PFAM_RIGHT: Denotes pfam id and location for reconstructed fusion transcript right to the breakpoint
\item Sample: Sample ids for the input fusion calls
\item Caller
}
}
\description{
detailed desc
1527 * 27
}
\references{
Star-Fusion documentation: https://github.com/STAR-Fusion/STAR-Fusion/wiki. TCGA sample information from supplementary file from Driver Fusions and Their Implications in the Development and Treatment of Human Cancers available at https://www.sciencedirect.com/science/article/pii/S2211124718303954
}
\keyword{datasets}
