% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{pfamDataBioMart}
\alias{pfamDataBioMart}
\title{pfamDataBioMart.RDS}
\format{
When read in via readRDS, the following information is available:
\itemize{
\item hgnc_symbol: Denotes gene symbol
\item pfam_id: Denotes pfam id
\item chromosome_name: Chromosome name for the gene region
\item gene_start: Genomic start coordinate of gene
\item gene_end: Genomic end coordinate of gene
\item strand: Strand information of the gene
\item NAME: Pfam domian name
\item DESC: Pfam domain description
\item domain_chr: Pfam domain chromsome name
\item domain_start: Pfam domain genomic start coordinate
\item domain_end: Pfam domain genomic end coordinate
}
}
\description{
detailed desc
57406 * 11
}
\references{
if any
}
\keyword{datasets}
