% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{sample_aliquot_tcga}
\alias{sample_aliquot_tcga}
\title{sample_aliquot_tcga.tsv}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item name: TCGA sample name
\item sample_id: TCGA sample id
\item aliquot_id: TCGA aliquot id
}
}
\description{
detailed desc
}
\references{
Supplementary file from Driver Fusions and Their Implications in the Development and Treatment of Human Cancers available at https://www.sciencedirect.com/science/article/pii/S2211124718303954
}
\keyword{datasets}
