% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{starfusion_example}
\alias{starfusion_example}
\title{starfusion_example.tsv}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item X.FusionName
\item JunctionReadCount
\item SpanningFragCount
\item SpliceType
\item LeftGene
\item LeftBreakpoint
\item RightGene
\item RightBreakpoint
\item LargeAnchorSupport
\item FFPM
\item LeftBreakDinuc
\item LeftBreakEntropy
\item RightBreakDinuc
\item RightBreakEntropy
\item annots
\item CDS_LEFT_ID
\item CDS_LEFT_RANGE
\item CDS_RIGHT_ID
\item CDS_RIGHT_RANGE
\item PROT_FUSION_TYPE
\item FUSION_MODEL
\item FUSION_CDS
\item FUSION_TRANSL
\item PFAM_LEFT
\item PFAM_RIGHT
\item tumor_id
}
}
\description{
detailed desc
One row of the typical output file from STAR-fusion (version TODO)
}
\references{
if any
}
\keyword{datasets}
