% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{arriba_example}
\alias{arriba_example}
\title{arriba_example.tsv}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item X.gene1: TODO - TODO: check if this is indeed the way it is supposed to be read in (if not, we would need to add a dependency to the data package as well)
\item gene2: TODO
\item strand1.gene.fusion.: TODO
\item strand2.gene.fusion.: TODO
\item breakpoint1: TODO
\item breakpoint2: TODO
\item site1: TODO
\item site2: TODO
\item type: TODO
\item direction1: TODO
\item direction2: TODO
\item split_reads1: TODO
\item split_reads2: TODO
\item discordant_mates: TODO
\item coverage1: TODO
\item coverage2: TODO
\item confidence: TODO
\item closest_genomic_breakpoint1: TODO
\item closest_genomic_breakpoint2: TODO
\item filters: TODO
\item fusion_transcript: TODO
\item reading_frame: TODO
\item peptide_sequence: TODO
\item read_identifiers: TODO
\item tumor_id: TODO
\item gene1..gene2: TODO
\item annots: TODO
}
}
\description{
detailed desc: TODO
25 rows are included, with details about 27 different features (see details below).
}
\keyword{datasets}
