% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{exonsToPlot}
\alias{exonsToPlot}
\title{exonsToPlot.RDS}
\format{
When read in via readRDS, the following information is available:
}
\description{
detailed desc
}
\references{
if any
}
\keyword{datasets}
