% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{exonsToPlot}
\alias{exonsToPlot}
\title{exonsToPlot.RDS}
\format{
When read in via readRDS, the following information is available:
\itemize{
\item contig
\item type
\item start
\item end
\item strand
\item attributes
\item geneName
\item geneID
\item transcript
\item exonNumber
}
}
\description{
detailed desc
58347 observations are included, with details about 7 different features (
see details below).
}
\references{
if any
}
\keyword{datasets}
