% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{expr_collapsed}
\alias{expr_collapsed}
\title{expr_collapsed.rds}
\format{
When read in via readRDS, the following information is available:
\itemize{
\item gene_id
\item SRR1069188
\item SRR1070986
\item SRR1071289
\item SRR1071880
\item SRR1072178
}
}
\description{
detailed desc
58581 * 6 columns
}
\references{
if any
}
\keyword{datasets}
