% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{FilteredFusionAnnoFuse}
\alias{FilteredFusionAnnoFuse}
\title{FilteredFusionAnnoFuse.tsv}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item LeftBreakpoint: TODO
\item RightBreakpoint: TODO
\item FusionName: TODO
\item Sample: TODO
\item Caller: TODO
\item Fusion_Type: TODO
\item JunctionReadCount: TODO
\item SpanningFragCount: TODO
\item Confidence: TODO
\item annots: TODO
\item Gene1A: TODO
\item Gene2A: TODO
\item Gene1B: TODO
\item Gene2B: TODO
\item BreakpointLocation: TODO
\item SpanningDelta: TODO
\item reciprocal_exists: TODO
\item Gene1A_anno: TODO
\item Gene1B_anno: TODO
\item Gene2A_anno: TODO
\item Gene2B_anno: TODO
\item Fusion_anno: TODO
\item integrated_diagnosis: TODO
\item broad_histology: TODO
\item Kids_First_Participant_ID: TODO
}
}
\description{
The exemplary output from the annoFuse workflow, containing a list of
filtered fusions, stored as tab-separated text file.
22101 observations are included, with details about 25 different features (
see details below).
}
\details{
This file is obtained from Star-Fusion, arriba and the rsem expression (fpkm)
files from release-v16-20200320 of
\href{https://github.com/AlexsLemonade/OpenPBTA-analysis}{OpenPBTA}
}
\keyword{datasets}
