% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{final_fusion_mmc2}
\alias{final_fusion_mmc2}
\title{final_fusion_mmc2.txt}
\format{
When read in via read.delim, the following information is available:
}
\description{
detailed desc
}
\references{
if any
}
\keyword{datasets}
