% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{fusionreference}
\alias{fusionreference}
\title{fusionreference.txt}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item FusionName
\item file
\item type
}
}
\description{
detailed desc
17754 * 3
}
\keyword{datasets}
