% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{genelistreference}
\alias{genelistreference}
\title{genelistreference.txt}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item Gene_Symbol
\item type
\item file
}
}
\description{
detailed desc
4954 * 3 columns
}
\keyword{datasets}
