% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{gtex_collapsed}
\alias{gtex_collapsed}
\title{gtex_collapsed.rds}
\format{
When read in via read.delim, the following information is available:
\itemize{
\item gene_id
\item SRR1069188
\item SRR1070986
\item SRR1071289
\item SRR1071880
\item SRR1072178
}
}
\description{
detailed desc
58581*6
}
\references{
if any, GTEX?
}
\keyword{datasets}
