% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{merged_starfusion_tcga}
\alias{merged_starfusion_tcga}
\title{merged_starfusion_tcga.tsv}
\format{
When read in via read.delim, the following information is available:
}
\description{
detailed desc
}
\references{
if any
}
\keyword{datasets}
