% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{sample_aliquot_tcga}
\alias{sample_aliquot_tcga}
\title{sample_aliquot_tcga.tsv}
\format{
When read in via read.delim, the following information is available:

if it applies
}
\description{
detailed desc
}
\references{
if any
}
\keyword{datasets}
