% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseData-data.R
\docType{data}
\name{starfusion_example}
\alias{starfusion_example}
\title{starfusion_example.tsv}
\format{
When read in via read.delim, the following information is available:
}
\description{
detailed desc
}
\references{
if any
}
\keyword{datasets}
